Always call iterable.close() in middleware.
When self.iterable is assigned to iter(iterable), it's
possible that the original iterable and self.iterable are two
different objects, since the original iterable is not required
to return self when you call iter(iterable).

If self.iterable differs from the original iterable, then calling
self.iterable.close() is not the same as calling iterable.close().
In those cases, iterable.close() is never called.

To fix this, we save away the close() method from the original
iterable, so that we can call it during ClosingIterator.close().

Fixes #885.